Bug fixes.Added pageable,filters,sort etc

+++ backend/src/main/java/com/example/medisyncpro/model/dto/MedicalReportDto.java
@@ -27,5 +27,5 @@ public class MedicalReportDto {
     /*private String doctorEmail;*/
     private LocalDateTime appointmentDate;
     private List<ServiceDto> services;
-    private BigDecimal totalPrice;
+    private int totalPrice;
 }
